refactoring super prims
